Pick state_lvldone.h apart
