+++ SAUCE/gamestates/state_lvldone.h
@@ -1,134 +1,8 @@
-
 CODE_BANK_PUSH("XCD_BANK_06")
+
+void checkcointimer();
+void checkcoinproceed();
 void lvl_done_update();
-void mouse_and_cursor();
-
-#if LEVELSET != 'A'
-
-const unsigned char palette_Credits2[16]={ 0x11,0x0f,0x10,0x30,0x11,0x0f,0x2a,0x39,0x11,0x28,0x17,0x0f,0x11,0x0f,0x11,0x21 };
-
-const unsigned char playmain[339]={
-0x03,0xfe,0x03,0x3f,0xff,0x03,0x27,0x20,0x21,0x22,0x23,0x24,0x25,0x26,0x27,0x28,
-0x29,0x22,0x23,0x22,0x2b,0x2c,0x2c,0xff,0x03,0x0f,0x30,0x31,0x32,0x33,0x34,0x35,
-0x36,0x37,0x38,0x39,0x32,0x33,0x3a,0x3b,0x3c,0x3d,0xff,0x03,0x0f,0x40,0xff,0x42,
-0x43,0x44,0x45,0x46,0x47,0x48,0x49,0x42,0x43,0x4a,0x4b,0x4c,0x4c,0xff,0x03,0x4d,
-0xf4,0xe8,0xe5,0xff,0xe6,0xe1,0xed,0xe9,0xe4,0xe1,0xf3,0xe8,0xff,0xe2,0xff,0xf3,
-0xe9,0xe4,0xe5,0xf3,0xff,0x03,0x0b,0xe1,0xf2,0xe5,0xff,0xe5,0xf8,0xf4,0xf2,0xe1,
-0xff,0xec,0xe5,0xf6,0xe5,0xec,0xf3,0xff,0xf4,0xe8,0xe1,0xf4,0xff,0x03,0x0a,0xe3,
-0xef,0xf5,0xec,0xe4,0xff,0xee,0xef,0xf4,0xff,0xe6,0xe9,0xf4,0xff,0xe9,0xee,0xff,
-0xf4,0xe8,0xe5,0xff,0x03,0x0b,0xed,0xe1,0xe9,0xee,0xff,0xe7,0xe1,0xed,0xe5,0xff,
-0x03,0x56,0xf4,0xef,0xff,0xf0,0xec,0xe1,0xf9,0xff,0xf4,0xe8,0xe5,0xff,0xed,0xe1,
-0xe9,0xee,0xff,0xec,0xe5,0xf6,0xe5,0xec,0xf3,0xff,0x03,0x08,0xe1,0xee,0xe4,0xff,
-0xed,0xef,0xf2,0xe5,0xff,0xf0,0xec,0xe5,0xe1,0xf3,0xe5,0xff,0xf0,0xec,0xe1,0xf9,
-0xff,0x03,0x0b,0xf4,0xe8,0xe5,0xff,0xed,0xe1,0xe9,0xee,0xff,0xe7,0xe1,0xed,0xe5,
-0xff,0x03,0x52,0xf6,0xe9,0xf3,0xe9,0xf4,0xff,0xf4,0xe8,0xe5,0xff,0xe3,0xef,0xed,
-0xed,0xf5,0xee,0xe9,0xf4,0xf9,0xff,0xf4,0xe1,0xe2,0xff,0x03,0x08,0xf4,0xef,0xff,
-0xf0,0xec,0xe1,0xf9,0xff,0xf4,0xe8,0xe5,0xff,0xe2,0xff,0xf3,0xe9,0xe4,0xe5,0xf3,
-0xff,0x03,0x0c,0xec,0xe5,0xf6,0xe5,0xec,0xf3,0xff,0x03,0x5c,0xf0,0xf2,0xe5,0xf3,
-0xf3,0xff,0xe1,0xee,0xf9,0xff,0xeb,0xe5,0xf9,0xff,0x03,0x14,0xf4,0xef,0xff,0xf2,
-0xe5,0xf4,0xf5,0xf2,0xee,0xff,0x03,0x4b,0xfe,0x03,0x3f,0x00,0x00,0x50,0x03,0x03,
-0x00,0x03,0x02,0x04,0x05,0x03,0x03,0x01,0x00,0x80,0x00,0x03,0x0e,0x04,0x00,0x08,
-0x0a,0x0a,0x00,0x03,0x07,0x80,0x00,0x03,0x03,0x0a,0x03,0x02,0x02,0x00,0x03,0x08,
-0x00,0x03,0x00
-};
-
-void state_playmain(){
-	mmc3_disable_irq();
-	
-	ppu_off();
-	pal_bright(0);
-    
-	pal_bg(palette_Credits2);
-	
-
-
-	vram_adr(NAMETABLE_A);
-	vram_unrle(playmain);
-
-	// __asm__("LDA mmc3PRG1Bank \nPHA ");
-    // mmc3_set_prg_bank_1(0);
-    // vram_adr(NAMETABLE_A);
-	//    vram_unrle(dem_funnies);
-    // __asm__("PLA \n JSR %v ", mmc3_set_prg_bank_1);
-
-	oam_clear();
-	mmc3_set_8kb_chr(MENUBANK);
-	mmc3_set_2kb_chr_bank_0(0xFF);	
-	mmc3_set_2kb_chr_bank_1(MOUSEBANK);	
-	ppu_on_all();
-	music_update();
-	pal_fade_to(0,4);
-	tmp1 = 0;
-
-	/*	Incomplete code for reproducing what is being screamed into the Famicom microphone
-	(He forgor about the PCM bit)
-	OG by UserSniper, commit 7e47f425
-	POKE(0x4015, 0b00010000);
-	while (1) {
-		POKE(0x4011, fc_mic_poll()<<4);
-	}
-	*/
-
-	
-	do {
-		oam_clear();
-		mouse_and_cursor();
-		music_update();
-		ppu_wait_nmi();
-		newrand();
-		kandoframecnt++;
-	} while (!joypad1.press && !mouse.left_press);
-	gameState = 1;
-	return;
-}	
-
-#endif
-
-const uint8_t difficulty_pal_A[] ={
-	0x21,	// easy
-	0x2A,	// normal
-	0x28,	// hard
-	0x16,	// harder
-	0x24,	// insane
-	0x16,	// demon
-#if !__VS_SYSTEM
-	0x28,	// auto
-#else
-	0x13,	// danger
-#endif
-};
-
-const uint8_t difficulty_pal_B[] ={
-	0x06,	// easy
-	0x30,	// normal
-	0x30,	// hard
-	0x30,	// harder
-	0x06,	// insane
-	0x30,	// demon
-#if !__VS_SYSTEM
-	0x0F,	// auto
-#else
-	0x21,	// danger
-#endif
-};
-
-void refreshmenu();
-void refreshmenu_part2();
-void code_checker();
-void check_if_music_stopped();
-void set_fun_settings();
-void checkcointimer(){
-	if (tmp1 == 1){
-		sfx_play(sfx_coin,0);
-		tmp1 = 50;
-	}
-}
-void checkcoinproceed(){
-	if (tmp1 == 0 || tmp1 == 30) {
-		tmp2++;
-		tmp1 = 1;
-	}
-}
 
 void state_lvldone() {
 	#define current_state tmp2
@@ -507,537 +381,6 @@ void state_lvldone() {
 	#undef top_scroll
 }
 
-
-
-CODE_BANK_PUSH("XCD_BANK_05")
-
-const unsigned char bgmtestscreen[];
-
-#if !__VS_SYSTEM
-#include "charmaps/bgm_charmap.h"
-#endif
-#include "music_soundTestTables.h"
-#include "sfx_soundTestTables.h"
-
-const char TEXT_debug_mode[] = "DEBUG MODE ENABLED";
-
-void bgmtest() {
-	song = 0;
-	temptemp6 = 0; 	
-	#define sfx tmp4
-	sfx = 0;
-	settingvalue = 0;
-  	famistudio_music_stop();
-  	music_update();
-	menuMusicCurrentlyPlaying=0;
-	pal_fade_to_withmusic(4,0);
-	mmc3_disable_irq();
-	ppu_off();
-	pal_bg(paletteMenu);
-	vram_adr(NAMETABLE_A);
-	vram_unrle(bgmtestscreen);   	
-	ppu_on_all();
-	pal_fade_to_withmusic(0,4);
-	while (1) {
-		
-		ppu_wait_nmi();
-		music_update();
-		oam_clear();
-		crossPRGBankJump0(check_if_music_stopped);
-		mouse_and_cursor();
-		 // read the first controller
-		kandoframecnt++;
-		if (kandoframecnt & 1 && mouse_timer) mouse_timer--;	
-		if (mouse.left_press) {
-			if ((mouse.x >= 0x63 && mouse.x <= 0x8C)) {
-				if (mouse.y >= 0x34 && mouse.y <= 0x3A) {		
-					settingvalue = 0;
-				}
-				else if ((mouse.y >= 0x6C && mouse.y <= 0x73)) {		
-					settingvalue = 1;
-				}
-			}
-			if ((mouse.x >= 0x2E && mouse.x <= 0xCC)) {
-				if ((mouse.y >= 0xB4 && mouse.y <= 0xBB)) {		
-					if (settingvalue == 1) { sfx_play(sfx, 0); }
-					else {
-						if (!temptemp6) { music_play(xbgmlookuptable[song]); temptemp6 = 1; songplaying = 1; }
-						else { famistudio_music_stop(); music_update(); temptemp6 = 0; songplaying = 0; }
-					}
-				}
-				if ((mouse.y >= 0xA4 && mouse.y <= 0xAB)) {
-					menu_music = song; sfx_play(sfx_orbs, 0);
-				}
-			}
-			if ((mouse.x >= 0x56 && mouse.x <= 0xA5) && (mouse.y >= 0x24 && mouse.y <= 0x2B)) {		
-				code_checker();
-				if (gameState == 0xF0) return;
-			}
-			if ((mouse.y >= 0x4E && mouse.y <= 0x5C)) {
-				if ((mouse.x >= 0x24 && mouse.x <= 0x2C)) {		
-					if (song == 0) {song = song_max - 1;} else song--; temptemp6 = 0; settingvalue = 0;
-				}
-				else if ((mouse.x >= 0xCC && mouse.x <= 0xD4)) {		
-					temptemp6 = 0; song++; if (song == song_max) {song = 0;} settingvalue = 0;
-				}
-			}
-			if ((mouse.y >= 0x86 && mouse.y <= 0x93)) {
-
-				if ((mouse.x >= 0x24 && mouse.x <= 0x2C)) {		
-					if (sfx == 0) {sfx = sfx_max - 1;} else sfx--; settingvalue = 1;
-				}
-				else if ((mouse.x >= 0xCC && mouse.x <= 0xD4)) {		
-					sfx++; if (sfx == sfx_max) {sfx= 0;} settingvalue = 1;
-				}
-			}
-			if ((mouse.x >= 0x35 && mouse.x <= 0xC4)) {
-				if (mouse.y >= 0xC4 && mouse.y <= 0xCD) {		
-					tmp3--;			
-					one_vram_buffer(' ', NTADR_A(11, 7));
-					one_vram_buffer(' ', NTADR_A(11, 14));
-					menuMusicCurrentlyPlaying = 1;
-					gameState = 1;
-					return;
-				}
-				else if (mouse.y >= 0x3D && mouse.y <= 0x64) {
-					if (!temptemp6) { music_play(xbgmlookuptable[song]); temptemp6 = 1; songplaying = 1; }
-					else { famistudio_music_stop(); music_update(); temptemp6 = 0; songplaying = 0; }
-				}
-				else if (mouse.y >= 0x75 && mouse.y <= 0x9C) {
-					sfx_play(sfx, 0);
-				}
-			}
-		}			
-	__A__ = idx16_load_hi_NOC(xbgmtextsUpper, song);
-	if (__A__) draw_padded_text(xbgmtextsUpper[song & 0x7F], xbgmtextsUpperSize[song], 18, NTADR_A(7, 10));
-	else one_vram_buffer_horz_repeat('$', 17, NTADR_A(7, 10));
-	__A__ = idx16_load_hi_NOC(xbgmtextsLower, song);
-	if (__A__) draw_padded_text(xbgmtextsLower[song & 0x7F], xbgmtextsLowerSize[song], 18, NTADR_A(7, 11));
-	else one_vram_buffer_horz_repeat('$', 17, NTADR_A(7, 11));
-	
-	draw_padded_text(sfxtexts[sfx & 0x7F], sfxtextSizes[sfx], 18, NTADR_A(7, 18));
-
-	
-		if (settingvalue == 0) {
-			one_vram_buffer('c', NTADR_A(11, 7));
-			one_vram_buffer(' ', NTADR_A(11, 14));
-			if (joypad1.press_right) { song++; temptemp6 = 0; if (song == song_max) {song = 0;} }
-			if (joypad1.press_left) { if (song == 0) {song = song_max - 1;} else song--; temptemp6 = 0; }
-			if (joypad1.press_a) {
-					if (!temptemp6) { music_play(xbgmlookuptable[song]); temptemp6 = 1; songplaying = 1; }
-					else { famistudio_music_stop(); music_update(); temptemp6 = 0; songplaying = 0; }
-			}					
-		}		
-		else if (settingvalue == 1) {
-			one_vram_buffer(' ', NTADR_A(11, 7));
-			one_vram_buffer('c', NTADR_A(11, 14));
-			if (joypad1.press_right) { sfx++; if (sfx == sfx_max) {sfx= 0;} };
-			if (joypad1.press_left) { if (sfx == 0) {sfx = sfx_max - 1;} else sfx--; }
-			if (joypad1.press_a) sfx_play(sfx, 0);
-		}
-
-		if (joypad1.press_down) settingvalue ^= 1;
-		if (joypad1.press_up) settingvalue ^= 1;
-		
-		if (joypad1.press_select) { menu_music = song; sfx_play(sfx_orbs, 0); }
-		if (joypad1.press_b) {
-			tmp3--;			
-			one_vram_buffer(' ', NTADR_A(11, 7));
-			one_vram_buffer(' ', NTADR_A(11, 14));
-			menuMusicCurrentlyPlaying = 1;
-			gameState = 1;
-			return;
-		}
-
-				
-		
-		// sound test codes
-		if (joypad1.press_start) {
-			code_checker();
-			if (gameState == 0xF0) return;
-		}
-	}
-}
-#define sfx tmp4
-void code_checker() {
-	last_gameState = gameState;
-	sfx_play(sfx_achievement_get, 0);
-	tmp3 = 1;
-
-	// bgm 9 & sfx 2
-	if (song == 0x9 && sfx == 0x2 && kandokidshack3 == 0) {
-		kandokidshack3++;
-		tmp3--;
-	}
-	
-	else if (song == 1 && sfx == 7 && kandokidshack3 == 1) {
-		all_levels_complete = 0xFC;
-		gameState = 0xF0; // fun settings gamestate
-		tmp3--;
-	}		
-	else kandokidshack3 = 0;
-	
-/*   debug code disabled
-	if (song == 0xB && sfx == 0x7) {
-		multi_vram_buffer_horz(TEXT_debug_mode, sizeof(TEXT_debug_mode)-1, NTADR_A(7,26));
-		options |= debugtoggle;
-		tmp3--;
-	}
-*/
-
-	// this is quite literally the greatest hack ever
-	// since sfx doesn't update until the next frame i can just
-	// overwrite the success sfx with the invalid one
-	if (tmp3) sfx_play(sfx_invalid, 0);	
-}
-
-#undef sfx
-
-
-const unsigned char bgmtestscreen[463]={
-0x01,0x02,0x01,0x04,0xae,0x02,0x01,0x13,0xae,0x02,0x01,0x09,0xaf,0x02,0x01,0x13,
-0xaf,0x02,0x01,0x06,0x06,0x07,0x04,0x01,0x17,0x06,0x07,0x02,0x01,0x03,0x08,0x09,
-0x05,0x01,0x17,0x08,0x09,0x02,0x01,0x03,0x0c,0xff,0x01,0x19,0x0d,0x02,0x01,0x03,
-0x0c,0xff,0x01,0x07,0xf3,0xef,0xf5,0xee,0xe4,0xff,0xf4,0xe5,0xf3,0xf4,0xff,0x01,
-0x07,0x0d,0x02,0x01,0x03,0x0c,0xff,0x01,0x19,0x0d,0x02,0x01,0x03,0x0c,0xff,0x01,
-0x09,0xed,0xf5,0xf3,0xe9,0xe3,0xff,0x01,0x0a,0x0d,0x02,0x01,0x03,0x0c,0xff,0x01,
-0x03,0x5c,0xfe,0x01,0x0f,0x5d,0xff,0x01,0x03,0x0d,0x02,0x01,0x03,0x0c,0xff,0x01,
-0x03,0xfe,0x01,0x11,0xff,0x01,0x03,0x0d,0x02,0x01,0x03,0x0c,0xff,0xff,0x6e,0xff,
-0xfe,0x01,0x11,0xff,0x6f,0xff,0xff,0x0d,0x02,0x01,0x03,0x0c,0xff,0xff,0x7e,0xff,
-0xfe,0x01,0x11,0xff,0x7f,0xff,0xff,0x0d,0x02,0x01,0x03,0x0c,0xff,0x01,0x03,0x6c,
-0xfe,0x01,0x0f,0x6d,0xff,0x01,0x03,0x0d,0x02,0x01,0x03,0x0c,0xff,0x01,0x19,0x0d,
-0x02,0x01,0x03,0x0c,0xff,0x01,0x09,0xf3,0xef,0xf5,0xee,0xe4,0xff,0x01,0x0a,0x0d,
-0x02,0x01,0x03,0x0c,0xff,0x01,0x03,0x5c,0xfe,0x01,0x0f,0x5d,0xff,0x01,0x03,0x0d,
-0x02,0x01,0x03,0x0c,0xff,0x01,0x03,0xfe,0x01,0x11,0xff,0x01,0x03,0x0d,0x02,0x01,
-0x03,0x0c,0xff,0xff,0x6e,0xff,0xfe,0x01,0x11,0xff,0x6f,0xff,0xff,0x0d,0x02,0x01,
-0x03,0x0c,0xff,0xff,0x7e,0xff,0xfe,0x01,0x11,0xff,0x7f,0xff,0xff,0x0d,0x02,0x01,
-0x03,0x0c,0xff,0x01,0x03,0x6c,0xfe,0x01,0x0f,0x6d,0xff,0x01,0x03,0x0d,0x02,0x01,
-0x03,0x0c,0xff,0x01,0x19,0x0d,0x02,0x01,0x03,0x0c,0xff,0xff,0xf3,0xe5,0xec,0xe5,
-0xe3,0xf4,0xe0,0xf3,0xe5,0xf4,0xff,0xed,0xe5,0xee,0xf5,0xff,0xed,0xf5,0xf3,0xe9,
-0xe3,0xff,0x01,0x02,0x0d,0x02,0x01,0x03,0x0c,0xff,0x01,0x19,0x0d,0x02,0x01,0x03,
-0x0c,0xff,0xff,0xf0,0xf2,0xe5,0xf3,0xf3,0xff,0xe1,0xff,0xf4,0xef,0xff,0xf0,0xec,
-0xe1,0xf9,0xff,0xe0,0xff,0xf3,0xf4,0xef,0xf0,0xff,0xff,0x0d,0x02,0x01,0x03,0x0c,
-0xff,0x01,0x19,0x0d,0x02,0x01,0x03,0x0c,0xff,0x01,0x03,0xf0,0xf2,0xe5,0xf3,0xf3,
-0xff,0xe2,0xff,0xf4,0xef,0xff,0xf2,0xe5,0xf4,0xf5,0xf2,0xee,0xff,0x01,0x04,0x0d,
-0x02,0x01,0x03,0x0a,0x0b,0x04,0x01,0x09,0x06,0x0e,0x0e,0x07,0x04,0x01,0x09,0x0a,
-0x0b,0x02,0x01,0x05,0x05,0x01,0x09,0x08,0x05,0x05,0x09,0x05,0x01,0x09,0x02,0x01,
-0x43,0xff,0x5c,0x5f,0x01,0x03,0x53,0xff,0x77,0x00,0x01,0x05,0xdd,0x77,0x00,0x01,
-0x05,0xdd,0x77,0x00,0x01,0x05,0xdd,0x77,0x00,0x01,0x05,0xdd,0x77,0x00,0x01,0x05,
-0xdd,0xf7,0x50,0x50,0xd0,0x70,0x50,0x50,0xfd,0x0f,0x01,0x06,0x0f,0x01,0x00
-};
-
-
-
-CODE_BANK_POP()
-
-
-void gameboy_check() {
-		if (gameboy_mode == 0) color_emphasis(COL_EMP_NORMAL);
-		else if (gameboy_mode == 1) color_emphasis(COL_EMP_GREY);
-		else if (gameboy_mode == 2) { color_emphasis(COL_EMP_GREYRED); }
-		else if (gameboy_mode == 3) { color_emphasis(COL_EMP_GREYGREEN); }
-		else if (gameboy_mode == 4) { color_emphasis(COL_EMP_GREYBLUE); }
-		else if (gameboy_mode == 5) { color_emphasis(COL_EMP_GREYYELLOW); }
-		else if (gameboy_mode == 6) { color_emphasis(COL_EMP_GREYPURPLE); }
-		else if (gameboy_mode == 7) { color_emphasis(COL_EMP_GREYCYAN); }
-		else if (gameboy_mode == 8) { color_emphasis(COL_EMP_GREYDARK); }
-}
-
-
-
-void colorinc() {
-	if (idx8_inc(icon_colors, settingvalue) & 0x30) {
-		if ((uint8_t)(icon_colors[settingvalue] & 0x0F) >= 0x0D)
-			idx8_store(icon_colors, settingvalue, (icon_colors[settingvalue] + 0x10) & 0x30);
-	} else {
-		if (((icon_colors[settingvalue] - 0x0D) & 0xFE) == 0)	// if color == 0x0D or 0x0E
-			icon_colors[settingvalue] = 0x0F;
-	}	
-}
-
-void colordec() {
-	if (idx8_dec(icon_colors, settingvalue) & 0x30) {
-		if ((uint8_t)(icon_colors[settingvalue] & 0x0F) >= 0x0D)
-			idx8_store(icon_colors, settingvalue, (icon_colors[settingvalue] & 0x30) | 0x0C);
-	} else {
-		if (((icon_colors[settingvalue] - 0x0D) & 0xFE) == 0)	// if color == 0x0D or 0x0E
-			icon_colors[settingvalue] = 0x0C;
-	}
-}	
-
-
-#include "charmaps/mainmenu_charmap.h"
-
-
-const unsigned char gameboytext1[]="  GREY";
-const unsigned char gameboytext2[]="   RED";
-const unsigned char gameboytext3[]=" GREEN";
-const unsigned char gameboytext4[]="  BLUE";
-const unsigned char gameboytext5[]="YELLOW";
-const unsigned char gameboytext6[]="PURPLE";
-const unsigned char gameboytext7[]="  CYAN";
-const unsigned char gameboytext8[]="  DARK";
-
-const unsigned char* const gameboytexts[] = {
-	0, gameboytext1, gameboytext2, gameboytext3, gameboytext4, gameboytext5, gameboytext6, gameboytext7, gameboytext8
-};
-
-const unsigned char gameboy_text_size[] = {
-	0,
-	sizeof(gameboytext1) - 1,
-	sizeof(gameboytext2) - 1,
-	sizeof(gameboytext3) - 1,
-	sizeof(gameboytext4) - 1,
-	sizeof(gameboytext5) - 1,
-	sizeof(gameboytext6) - 1,
-	sizeof(gameboytext7) - 1,
-	sizeof(gameboytext8) - 1
-};
-
-
-
-void funsettings() {
-	settingvalue = 0; 
-	pal_fade_to_withmusic(4,0);
-	ppu_off();
-	pal_bg(paletteMenu);
-	vram_adr(NAMETABLE_A);
-	vram_unrle(funsettingscreen);   
-	//kandotemp4 = 0;
-	mmc3_set_2kb_chr_bank_0(0xFF);
-	mmc3_set_2kb_chr_bank_1(MOUSEBANK);
-	ppu_on_all();
-	one_vram_buffer('c', NTADR_A(4, 5));	// settingvalue is set to 0 by default	
-	pal_fade_to_withmusic(0,4);
-	while (1) {
-		ppu_wait_nmi();
-		music_update();
-		oam_clear();
-		crossPRGBankJump0(check_if_music_stopped);
-		mouse_and_cursor();
-		 // read the first controller
-		if (mouse.left_press) {
-			if (mouse.x >= 0x2D && mouse.x <= 0xDD) {
-				if (mouse.y >= 0x24 && mouse.y <= 0x2C) {
-					settingvalue = 0; set_fun_settings();
-				}
-				else if (mouse.y >= 0x34 && mouse.y <= 0x3C) {
-					settingvalue = 1; set_fun_settings();
-				}
-				else if (mouse.y >= 0x44 && mouse.y <= 0x4C) {
-					settingvalue = 2; set_fun_settings();
-				}
-				else if (mouse.y >= 0x54 && mouse.y <= 0x5C) {
-					settingvalue = 3; set_fun_settings();
-				}
-				else if (mouse.y >= 0x64 && mouse.y <= 0x6C) {
-					settingvalue = 4; set_fun_settings();
-				}
-				else if (mouse.y >= 0x74 && mouse.y <= 0x7C) {
-					settingvalue = 5; set_fun_settings();
-				}
-				else if (mouse.y >= 0x84 && mouse.y <= 0x8C) {
-					settingvalue = 6; set_fun_settings();
-				}
-				else if (mouse.y >= 0x94 && mouse.y <= 0x9C) {
-					settingvalue = 7; set_fun_settings();
-				}
-				else if (mouse.y >= 0xA4 && mouse.y <= 0xAC) {
-					settingvalue = 8; set_fun_settings();
-				}
-
-			}
-			if ((mouse.x >= 0x1D && mouse.x <= 0xDD) && (mouse.y >= 0xBC && mouse.y <= 0xC4)) {		
-				gameState = last_gameState;
-				return;
-			}
-
-		}	
-		if (invisible) 	one_vram_buffer('g', NTADR_A(26, 5));	// believe it or not, 
-		else 	one_vram_buffer('f', NTADR_A(26, 5));	// this is auto optimized by cc65
-
-		if ((options & platformer)) 	one_vram_buffer('g', NTADR_A(26, 7));	// believe it or not, 
-		else 	one_vram_buffer('f', NTADR_A(26, 7));	// this is auto optimized by cc65
-
-		if (retro_mode) 	one_vram_buffer('g', NTADR_A(26, 9));	// believe it or not, 
-		else 	one_vram_buffer('f', NTADR_A(26, 9));	// this is auto optimized by cc65
-
-		if (discomode) 	one_vram_buffer('g', NTADR_A(26, 11));	// believe it or not, 
-		else 	one_vram_buffer('f', NTADR_A(26, 11));	// this is auto optimized by cc65
-		
-		if (invisblocks) 	one_vram_buffer('g', NTADR_A(26, 13));	// believe it or not, 
-		else 	one_vram_buffer('f', NTADR_A(26, 13));	// this is auto optimized by cc65
-		
-		if (cam_seesaw) 	one_vram_buffer('g', NTADR_A(26, 15));	// believe it or not, 
-		else 	one_vram_buffer('f', NTADR_A(26, 15));	// this is auto optimized by cc65
-
-		if (cursedmusic & 0x40) 	one_vram_buffer('g', NTADR_A(26, 19));	// believe it or not, 
-		else 	one_vram_buffer('f', NTADR_A(26, 19));	// this is auto optimized by cc65
-		
-		if (practice_music_sync) 	one_vram_buffer('g', NTADR_A(26, 21));	// believe it or not, 
-		else 	one_vram_buffer('f', NTADR_A(26, 21));	// this is auto optimized by cc65
-
-
-//		if (cursedmusic) 	one_vram_buffer('g', NTADR_A(26, 21));	// believe it or not, 
-//		else 	one_vram_buffer('f', NTADR_A(26, 21));	// this is auto optimized by cc65
-		
-		__A__ = idx16_load_hi_NOC(gameboytexts, gameboy_mode);
-		if (__A__) { draw_padded_text(gameboytexts[gameboy_mode & 0x7F], gameboy_text_size[gameboy_mode], 8, NTADR_A(19, 17)); 	one_vram_buffer('g', NTADR_A(26, 17));	}// believe it or not, 
-		else { one_vram_buffer_horz_repeat('$', 8, NTADR_A(19, 17)); one_vram_buffer('f', NTADR_A(26, 17)); }
-
-
-
-		tmp1 = settingvalue;
-
-		if (joypad1.press & (PAD_RIGHT | PAD_DOWN)) {
-			if (settingvalue == 8) { settingvalue = 0; }
-			else { settingvalue++;  }
-		}
-
-		if (joypad1.press & (PAD_LEFT | PAD_UP)) {
-			if (settingvalue == 0) { settingvalue = 8; }
-			else { settingvalue--;  }
-		}
-
-		if (tmp1 != settingvalue) {
-			// NTADR_A = (NAMETABLE_A|(((y)<<5)|(x)))
-			// (tmp1 * 2) << 5 = tmp1<<6 = (tmp1<<8)>>2
-			one_vram_buffer(' ', NTADR_A(4, 5)+((tmp1<<8)>>2));
-			one_vram_buffer('c', NTADR_A(4, 5)+((settingvalue<<8)>>2));
-		}
-
-		one_vram_buffer('X'-0x1B, NTADR_A(26, 11));
-		if (discomode & 0x02) { one_vram_buffer('2' - 0x20, NTADR_A(25, 11)); }
-		else if (discomode & 0x04) { one_vram_buffer('3' - 0x20, NTADR_A(25, 11));}
-		else if (discomode & 0x08) { one_vram_buffer('4' - 0x20, NTADR_A(25, 11));}
-		else if (discomode & 0x10) { one_vram_buffer('5' - 0x20, NTADR_A(25, 11));}
-		else if (discomode & 0x01) { one_vram_buffer('1' - 0x20, NTADR_A(25, 11));}
-		else { one_vram_buffer(' '-0x01, NTADR_A(25, 11)); one_vram_buffer(0x0F, NTADR_A(26, 11)); }
-
-		if (joypad1.press & (PAD_START | PAD_A)) {
-			set_fun_settings();
-		}
-			
-		if (joypad1.press_b) {
-			gameState = last_gameState;
-			return;
-		}
-
-		crossPRGBankJump0(gameboy_check);
-
-//		if (gameboy_mode) kandotemp4 = 1;
-		kandoframecnt++;
-		if (kandoframecnt & 1 && mouse_timer) mouse_timer--;	
-						
-	}
-}
-
-void set_fun_settings() {
-	switch (settingvalue) {
-		case 0x00: invisible ^= 1; break;
-		case 0x01: options ^= platformer; break;
-		case 0x02: retro_mode ^= 1; break;
-		case 0x03: 
-			if (discomode & 0x10) discomode = 0;
-			else {
-				discomode = discomode << 1;
-				if (discomode == 0) ++discomode;
-			}
-			
-			break;
-		case 0x04: invisblocks ^= 1; break;
-		case 0x05: cam_seesaw = (cam_seesaw > 0 ? 0 : 1); break;
-		case 0x06: gameboy_mode = (gameboy_mode == 8 ? 0 : gameboy_mode + 1); break;
-		case 0x07: cursedmusic ^= 0x40; break;
-		case 0x08: practice_music_sync ^= 0x01; break;
-	};
-}	
-
-#include "charmaps/color1_charmap.h"
-#include "menutext.h"
-
-const char coin_counter[][3] = {
-  "___",
-  "^__",
-  "_^_",
-  "^^_",
-  "__^",
-  "^_^",
-  "_^^",
-  "^^^",
-};
-
-/*
-	Refreshes level name & number
-*/
-void refreshmenu() {
-	tmp5 = ((level&1)<<10);
-	set_scroll_x(((level-tmp4)&1)<<8);
-	
-	__A__ = idx16_load_hi_NOC(levelTextsUpper, level);
-	if (__A__) draw_padded_text(levelTextsUpper[level & 0x7F], levelTextsUpperSize[level], 17, NTADR_A(8, 10)+tmp5);
-	else one_vram_buffer_horz_repeat(' ', 17, NTADR_A(8, 10)+tmp5);
-	// if (leveltexts2[level]) // always true
-	draw_padded_text(levelTextsLower[level & 0x7F], levelTextsLowerSize[level], 17, NTADR_A(8, 11)+tmp5);
-
-	if (LEVELCOMPLETE[level]) { one_vram_buffer('y', NTADR_A(7, 9)+tmp5);
-	one_vram_buffer('z', NTADR_A(8, 9)+tmp5); }
-	else one_vram_buffer_horz_repeat(' ', 2, NTADR_A(7, 9)+tmp5);
-
-	{	// write the difficulty
-		tmp1 = difficulty_list[level];
-		pal_col(0x0a, difficulty_pal_A[tmp1]);
-		pal_col(0x0b, difficulty_pal_B[tmp1]);
-		pal_set_update();
-		
-		tmp1 = (tmp1 << 1) + 'a';
-		one_vram_buffer(tmp1, NTADR_A(7, 10)+tmp5);
-		one_vram_buffer(++tmp1, NTADR_A(8, 10)+tmp5);
-		one_vram_buffer((tmp1 += ('c'-'b')), NTADR_A(7, 11)+tmp5);
-		one_vram_buffer(++tmp1, NTADR_A(8, 11)+tmp5);
-		
-
-	}
-	// Star count stuff
-		printDecimal(stars_list[level], 2, '0', ' ', NTADR_A(22, 9)+tmp5);
-
-	// level number
-	#ifdef FLAG_ENABLE_TEST_LEVELS
-		printDecimal(level, 3, '0', ' ', NTADR_A(29, 2));
-		printDecimal(level, 3, '0', ' ', NTADR_B(29, 2));
-	#endif
-};
-
-void refreshmenu_part2() {
-	
-	// Normal level completeness stuff
-		//printDecimal(level_completeness_normal[level], 3, '0', ' ', NTADR_A(14, 16)+tmp5);
-
-	// Practice level completeness stuff
-		//printDecimal(level_completeness_practice[level], 3, '0', ' ', NTADR_A(14, 19)+tmp5);
-
-	//palette stuff
-		tmp3 = level % 9;
-		pal_col(0,colors_list[tmp3]);
-		pal_col(0xE,colors_list[tmp3]);
-		
-		pal_spr(paletteLVLSelectSP);
-		pal_col(0x10,colors_list[tmp3]);
-		pal_col(0x1E,colors_list[tmp3]);
-		pal_set_update();
-	//coin stuff
-		coins = 0;
-
-
-	// then in the function...
-	// combine all three into a single number from 0 - 7 to represent which coins have been grabbed
-		tmp7 = byte((byte(coin3_obtained[level] << 1) | coin2_obtained[level]) << 1) | coin1_obtained[level];
-		tmp7 = byte(tmp7<<1) + tmp7;
-	// actually draw the coins
-		multi_vram_buffer_horz((const char * const)coin_counter+tmp7, 3, NTADR_A(22, 12)+tmp5);
-
-}
-
 #include "charmaps/endlevel_charmap.h"
 void lvl_done_update() {
 	if (menuselection) {
@@ -1054,83 +397,17 @@ void lvl_done_update() {
 }	
 
 
-#define BAR_TILES 20
-#define CENTER_TILES BAR_TILES - 2
-#define BAR_MAX 100
-#define NUMBER_PER_TILES BAR_MAX / BAR_TILES
-void draw_progress_bar() {
-	#define nametable_A !(level & 1)
-	
-	drawBarFlag--;
-
-	if (tmp7 >= NUMBER_PER_TILES) {
-		if (nametable_A) {
-			one_vram_buffer(0x8c, NTADR_A(tmp1, tmp2));
-		} else {
-			one_vram_buffer(0x8c, NTADR_B(tmp1, tmp2));
-		}
-		high_byte(tmp6) = (tmp7 >= BAR_MAX ? BAR_MAX - NUMBER_PER_TILES : tmp7);
-
-		
-		high_byte(tmpA) = 0;
-		do {
-			high_byte(tmp6) -= NUMBER_PER_TILES;
-			high_byte(tmpA) += 1;
-		} while (high_byte(tmp6) >= NUMBER_PER_TILES);
-		
-		if (nametable_A) {
-			one_vram_buffer_horz_repeat(0x6b, high_byte(tmpA), NTADR_A(tmp1 + 1, tmp2));
-			one_vram_buffer_horz_repeat(0x02, BAR_TILES - high_byte(tmpA), NTADR_A(tmp1 + high_byte(tmpA), tmp2));
-		} else {
-			one_vram_buffer_horz_repeat(0x6b, high_byte(tmpA), NTADR_B(tmp1 + 1, tmp2));
-			one_vram_buffer_horz_repeat(0x02, BAR_TILES - high_byte(tmpA), NTADR_B(tmp1 + high_byte(tmpA), tmp2));
-		}
-	} else {
-		if (nametable_A) {
-			one_vram_buffer(0x7c, NTADR_A(tmp1, tmp2));
-			one_vram_buffer_horz_repeat(0x02, CENTER_TILES, NTADR_A(tmp1 + 1, tmp2));
-		} else {
-			one_vram_buffer(0x7c, NTADR_B(tmp1, tmp2));
-			one_vram_buffer_horz_repeat(0x02, CENTER_TILES, NTADR_B(tmp1 + 1, tmp2));
-		}
+void checkcointimer(){
+	if (tmp1 == 1){
+		sfx_play(sfx_coin,0);
+		tmp1 = 50;
 	}
-	
-	if (tmp7 >= BAR_MAX) {
-		if (nametable_A) {
-			one_vram_buffer(0x8d, NTADR_A(tmp1 + (BAR_TILES - 1), tmp2));
-		} else {
-			one_vram_buffer(0x8d, NTADR_B(tmp1 + (BAR_TILES - 1), tmp2));
-		}
-	} else {
-		if (nametable_A) {
-			one_vram_buffer(0x7d, NTADR_A(tmp1 + (BAR_TILES - 1), tmp2));
-		} else {
-			one_vram_buffer(0x7d, NTADR_B(tmp1 + (BAR_TILES - 1), tmp2));
-		}
-	}
-	#undef nametable_A
 }
-
-#define TOTAL_PIXELS 8 * BAR_TILES // 8 * 20 = 160
-#define PIXELS_PER_UNIT (TOTAL_PIXELS * 10) / (BAR_MAX) // (1600 / 100) = 1.6 | multiplied by 10 because of no floats
-#define PIXELS_PER_PERCENTAGE (PIXELS_PER_UNIT * 256) / 10 // 1.6 * 256 = 409.6
-
-void calculate_sprite_pos() {
-	high_byte(tmp5) = tmp1;
-	low_byte(tmp5) = 0;
-	tmp3 = 1;
-
-	if (tmp7 != 0) {
-		do {
-			tmp3++;
-			tmp5 += PIXELS_PER_PERCENTAGE;
-		} while (tmp3 < tmp7);
+void checkcoinproceed(){
+	if (tmp1 == 0 || tmp1 == 30) {
+		tmp2++;
+		tmp1 = 1;
 	}
-
-	tmp1 = high_byte(tmp5);
 }
-
-
 
 CODE_BANK_POP()
-
