+++ SAUCE/menustates/misc_xcd_bank_06.c
@@ -1,0 +1,35 @@
+// Will be table'd and possibly even inlined
+
+void gameboy_check() {
+		if (gameboy_mode == 0) color_emphasis(COL_EMP_NORMAL);
+		else if (gameboy_mode == 1) color_emphasis(COL_EMP_GREY);
+		else if (gameboy_mode == 2) { color_emphasis(COL_EMP_GREYRED); }
+		else if (gameboy_mode == 3) { color_emphasis(COL_EMP_GREYGREEN); }
+		else if (gameboy_mode == 4) { color_emphasis(COL_EMP_GREYBLUE); }
+		else if (gameboy_mode == 5) { color_emphasis(COL_EMP_GREYYELLOW); }
+		else if (gameboy_mode == 6) { color_emphasis(COL_EMP_GREYPURPLE); }
+		else if (gameboy_mode == 7) { color_emphasis(COL_EMP_GREYCYAN); }
+		else if (gameboy_mode == 8) { color_emphasis(COL_EMP_GREYDARK); }
+}
+
+// Will be merged into customize_screen.c
+
+void colorinc() {
+	if (idx8_inc(icon_colors, settingvalue) & 0x30) {
+		if ((uint8_t)(icon_colors[settingvalue] & 0x0F) >= 0x0D)
+			idx8_store(icon_colors, settingvalue, (icon_colors[settingvalue] + 0x10) & 0x30);
+	} else {
+		if (((icon_colors[settingvalue] - 0x0D) & 0xFE) == 0)	// if color == 0x0D or 0x0E
+			icon_colors[settingvalue] = 0x0F;
+	}	
+}
+
+void colordec() {
+	if (idx8_dec(icon_colors, settingvalue) & 0x30) {
+		if ((uint8_t)(icon_colors[settingvalue] & 0x0F) >= 0x0D)
+			idx8_store(icon_colors, settingvalue, (icon_colors[settingvalue] & 0x30) | 0x0C);
+	} else {
+		if (((icon_colors[settingvalue] - 0x0D) & 0xFE) == 0)	// if color == 0x0D or 0x0E
+			icon_colors[settingvalue] = 0x0C;
+	}
+}
