+++ SAUCE/menustates/playmain.c
@@ -1,0 +1,67 @@
+const unsigned char palette_Credits2[];
+const unsigned char playmain[];
+
+void state_playmain(){
+	mmc3_disable_irq();
+	
+	ppu_off();
+	pal_bright(0);
+    
+	pal_bg(palette_Credits2);
+	
+	vram_adr(NAMETABLE_A);
+	vram_unrle(playmain);
+
+	// __asm__("LDA mmc3PRG1Bank \nPHA ");
+    // mmc3_set_prg_bank_1(0);
+    // vram_adr(NAMETABLE_A);
+	//    vram_unrle(dem_funnies);
+    // __asm__("PLA \n JSR %v ", mmc3_set_prg_bank_1);
+
+	oam_clear();
+	mmc3_set_8kb_chr(MENUBANK);
+	mmc3_set_2kb_chr_bank_0(0xFF);	
+	mmc3_set_2kb_chr_bank_1(MOUSEBANK);	
+	ppu_on_all();
+	music_update();
+	pal_fade_to(0,4);
+	tmp1 = 0;
+
+	do {
+		oam_clear();
+		mouse_and_cursor();
+		music_update();
+		ppu_wait_nmi();
+		newrand();
+		kandoframecnt++;
+	} while (!joypad1.press && !mouse.left_press);
+	gameState = 1;
+	return;
+}	
+
+const unsigned char palette_Credits2[16]={ 0x11,0x0f,0x10,0x30,0x11,0x0f,0x2a,0x39,0x11,0x28,0x17,0x0f,0x11,0x0f,0x11,0x21 };
+
+const unsigned char playmain[339]={
+0x03,0xfe,0x03,0x3f,0xff,0x03,0x27,0x20,0x21,0x22,0x23,0x24,0x25,0x26,0x27,0x28,
+0x29,0x22,0x23,0x22,0x2b,0x2c,0x2c,0xff,0x03,0x0f,0x30,0x31,0x32,0x33,0x34,0x35,
+0x36,0x37,0x38,0x39,0x32,0x33,0x3a,0x3b,0x3c,0x3d,0xff,0x03,0x0f,0x40,0xff,0x42,
+0x43,0x44,0x45,0x46,0x47,0x48,0x49,0x42,0x43,0x4a,0x4b,0x4c,0x4c,0xff,0x03,0x4d,
+0xf4,0xe8,0xe5,0xff,0xe6,0xe1,0xed,0xe9,0xe4,0xe1,0xf3,0xe8,0xff,0xe2,0xff,0xf3,
+0xe9,0xe4,0xe5,0xf3,0xff,0x03,0x0b,0xe1,0xf2,0xe5,0xff,0xe5,0xf8,0xf4,0xf2,0xe1,
+0xff,0xec,0xe5,0xf6,0xe5,0xec,0xf3,0xff,0xf4,0xe8,0xe1,0xf4,0xff,0x03,0x0a,0xe3,
+0xef,0xf5,0xec,0xe4,0xff,0xee,0xef,0xf4,0xff,0xe6,0xe9,0xf4,0xff,0xe9,0xee,0xff,
+0xf4,0xe8,0xe5,0xff,0x03,0x0b,0xed,0xe1,0xe9,0xee,0xff,0xe7,0xe1,0xed,0xe5,0xff,
+0x03,0x56,0xf4,0xef,0xff,0xf0,0xec,0xe1,0xf9,0xff,0xf4,0xe8,0xe5,0xff,0xed,0xe1,
+0xe9,0xee,0xff,0xec,0xe5,0xf6,0xe5,0xec,0xf3,0xff,0x03,0x08,0xe1,0xee,0xe4,0xff,
+0xed,0xef,0xf2,0xe5,0xff,0xf0,0xec,0xe5,0xe1,0xf3,0xe5,0xff,0xf0,0xec,0xe1,0xf9,
+0xff,0x03,0x0b,0xf4,0xe8,0xe5,0xff,0xed,0xe1,0xe9,0xee,0xff,0xe7,0xe1,0xed,0xe5,
+0xff,0x03,0x52,0xf6,0xe9,0xf3,0xe9,0xf4,0xff,0xf4,0xe8,0xe5,0xff,0xe3,0xef,0xed,
+0xed,0xf5,0xee,0xe9,0xf4,0xf9,0xff,0xf4,0xe1,0xe2,0xff,0x03,0x08,0xf4,0xef,0xff,
+0xf0,0xec,0xe1,0xf9,0xff,0xf4,0xe8,0xe5,0xff,0xe2,0xff,0xf3,0xe9,0xe4,0xe5,0xf3,
+0xff,0x03,0x0c,0xec,0xe5,0xf6,0xe5,0xec,0xf3,0xff,0x03,0x5c,0xf0,0xf2,0xe5,0xf3,
+0xf3,0xff,0xe1,0xee,0xf9,0xff,0xeb,0xe5,0xf9,0xff,0x03,0x14,0xf4,0xef,0xff,0xf2,
+0xe5,0xf4,0xf5,0xf2,0xee,0xff,0x03,0x4b,0xfe,0x03,0x3f,0x00,0x00,0x50,0x03,0x03,
+0x00,0x03,0x02,0x04,0x05,0x03,0x03,0x01,0x00,0x80,0x00,0x03,0x0e,0x04,0x00,0x08,
+0x0a,0x0a,0x00,0x03,0x07,0x80,0x00,0x03,0x03,0x0a,0x03,0x02,0x02,0x00,0x03,0x08,
+0x00,0x03,0x00
+};
